+++ sdk/mq/callback/network.js
@@ -139,7 +139,6 @@ class NetworkMgr extends EventEmitter {
     this.#mgrMsgHandler = handler;
 
     this.#uldata.on(Events.Status, this.#gmqStatusHandler.bind(this));
-    this.#uldata.setMsgHandler(this.#gmqMsgHandler.bind(this));
     this.#uldata.connect();
 
     this.#dldata.on(Events.Status, this.#gmqStatusHandler.bind(this));
@@ -147,7 +146,6 @@ class NetworkMgr extends EventEmitter {
     this.#dldata.connect();
 
     this.#dldataResult.on(Events.Status, this.#gmqStatusHandler.bind(this));
-    this.#dldataResult.setMsgHandler(this.#gmqMsgHandler.bind(this));
     this.#dldataResult.connect();
 
     this.#ctrl.on(Events.Status, this.#gmqStatusHandler.bind(this));
@@ -308,7 +306,7 @@ class NetworkMgr extends EventEmitter {
   /**
    * The handler for the gmq.Queue#status events.
    */
-  #gmqStatusHandler(_queue, _status) {
+  #gmqStatusHandler(_status) {
     let status;
     if (
       this.#uldata.status() === Status.Connected &&